creating component for manga registration form
